Move python preprocessing to adapters
