+++ model_api/python/openvino/model_api/adapters/__init__.py
@@ -18,7 +18,7 @@
 from .onnx_adapter import ONNXRuntimeAdapter
 from .openvino_adapter import OpenvinoAdapter, create_core, get_user_config
 from .ovms_adapter import OVMSAdapter
-from .utils import Layout
+from .utils import INTERPOLATION_TYPES, RESIZE_TYPES, InputTransform, Layout
 
 __all__ = [
     "create_core",
@@ -27,4 +27,7 @@ __all__ = [
     "OpenvinoAdapter",
     "OVMSAdapter",
     "ONNXRuntimeAdapter",
+    "RESIZE_TYPES",
+    "InputTransform",
+    "INTERPOLATION_TYPES",
 ]
